Implement simple database access in rsdroid
This is required as basic rust functionality will keep a locked connection open
on the database, so all database access needs to go through rslib-bridge.

This involves:
* Parameterised queries
* Nested transactions
* Column name mapping

This needs improving:

All data is sent in memory currently, which will likely OOM low memory devices.
We need to stream

We need to lock access to the rust backend when we are in a transaction
as rust only uses one connection to the database per Backend.

This means that transactions could theoretically interfere.

We need better exceptions from database problems - the correct exception types
are not thrown.

We need better separation of the database and the backend:
openAnkiDroidCollection takes the path to open the database,
it should be in RustSupportSQLiteDatabase

Closing the database also needs to be better