Use zip of desktop artifacts in build
Instead of checking buildinfo.txt and the web files into this repo,
we should either generate them as part of the build, or rely on a pinned
pre-generated file. The latter is the easier short-term solution, as it
requires minimal updates to CI, and won't balloon out the build times.
It's more work when bumping versions though, so in the long time AD may
want to consider adopting Bazel in this repo once this code has sufficiently
proven itself.

There's probably room for improvement in the download/unpack task - I'll
leave that for someone with more Gradle experience. I am not able to test
the artifact build/upload from the repo fork, so have temporarily placed the
generated file on ankiweb.