Fix incorrect timezone on some devices; bump version
The timezone dependency Anki uses recently changed the way they gather
timezone info. They used to rely on a libc call, which could cause crashes
if env vars happened to be modified in another thread at the same, so
they decided to reimplement the functionality in Rust instead to address
a CVE: https://github.com/chronotope/chrono/issues/602

Android appeared to be supported (https://github.com/chronotope/chrono/releases/tag/v0.4.21),
but it seems that this only applies to some devices, or was not tested.
On a Pixel 2/API 31 sim, the /system/usr/share/zoneinfo path it refers
to exists, but it does not have the individual timezone files that the
library expects to find.

Until a better solution can be found, this commit reverts to an earlier
version of the library where TZ lookup was working. It depends on
https://github.com/ankidroid/anki/pull/7