Update to the latest desktop
- Update the backend protobuf generation to match the refactoring done
on the desktop code recently. This involved switching from the Python-based
script to a new one in Rust. Comments from the proto files are now
included in the generated backend bindings.
- Switch from protobuf to protobuf-lite, which is what Google recommend
for use on Android.
- Recent desktop changes introduce a protobuf construct that protobuf-lite
doesn't support, so the offending file needed to be removed
as part of the build. This was easier to do by directly invoking protobuf,
instead of relying on the Gradle protobuf plugin. As a bonus, this should
also solve the protobuf-sources-are-missing-from-bundle issue. (file
removal is no longer required due to a refactor).
- Update Rust deps. is_terminal was failing to build on Windows, and
is no longer required after updating to the latest android_logger
- Rumo contributed changes to the upstream Chrono crate which make
it work properly on Android, and these changes have been merged in,
so AnkiDroid no longer needs to pin an older version.
- Update to latest NDK, as latest Rust requires it (closes #180).
- Update to latest Rust
- Add a script to make it easier to reformat Rust code when it fails
checks.