ignore 'checkMainThreadOp' unit test error
Unit tests are much faster if they don't depend on Robolectric,
the downside is that we can't depend on Android functionality

Looper.getMainLooper().thread is run only in `checkMainThreadOp`
and this is a logging function, so we can ignore the error

java.lang.RuntimeException: Method getMainLooper in android.os.Looper not mocked. See https://developer.android.com/r/studio-ui/build/not-mocked for details.
		at android.os.Looper.getMainLooper(Looper.java)